+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.2
+
+- Added an optional parameter `root`, a `bool` value that decides if the `SnackBar` should be shown in the context of the nearest or furthest `Scaffold`
+- Added an optional parameter `scaffoldState` that can be passed to specify a concrete `Scaffold` to show the `SnackBar` within
+
 ## 1.0.1+2
 
 - Fixed readme image